Get netns building on windows
